Bugs 2102

+++ Raspberry Pi/setup.sh
@@ -171,6 +171,11 @@ install_website ()
 	# piTransfer.py will add to this file the name of every image it successfully transfers
 	touch photos/uploadedOK.txt
 
+	if [ -f default_image.jpg ];
+	then
+		mv -fv default_image.jpg ~/photos/default_image.jpg
+	fi
+
 	chown -R pi:www-data /home/pi
 
 	if [ -f www/intvlm8r.old ];
@@ -679,9 +684,72 @@ END
 
 unmake_ap ()
 {
-	systemctl disable dnsmasq #Stops it launching on bootup
-	systemctl disable hostapd
-	sed -i -E "s|^\s*#*\s*(DAEMON_CONF=\")(.*\")|## \1\2|" /etc/default/hostapd # DOUBLE-Comment-out
+	if systemctl --all --type service | grep -q "dnsmasq";
+	then
+		systemctl disable dnsmasq #Stops it launching on bootup
+		echo 'Disabled dnsmasq'
+	fi
+	if systemctl --all --type service | grep -q "hostapd";
+	then
+		systemctl disable hostapd
+		echo 'Disabled hostapd'
+		sed -i -E "s|^\s*#*\s*(DAEMON_CONF=\")(.*\")|## \1\2|" /etc/default/hostapd # DOUBLE-Comment-out
+	fi
+
+	oldCountry=$(sed -n -E 's|^\s*country=(.*)$|\1|p' /etc/wpa_supplicant/wpa_supplicant.conf | tail -1) # Delimiter needs to be '|'
+	oldSsid=$(sed -n -E 's|^\s*ssid="(.*)"$|\1|p' /etc/wpa_supplicant/wpa_supplicant.conf | tail -1) # Delimiter needs to be '|'
+	oldPsk=$(sed -n -E 's|^\s*psk="(.*)"$|\1|p' /etc/wpa_supplicant/wpa_supplicant.conf | tail -1) # Delimiter needs to be '|'
+	while true; do
+		read -e -i "$oldCountry" -p "Set your two-letter WiFi country code : " newCountry
+		if [ -z "$newCountry" ];
+		then
+			echo -e "Error: Country value cannot be empty."
+			echo ''
+			continue
+		fi
+		break
+	done
+	while true; do
+		read -e -i "$oldSsid"    -p "Set the network's SSID                : " newSsid
+		if [ -z "$newSsid" ];
+		then
+			echo -e "Error: SSID value cannot be empty."
+			echo ''
+			continue
+		fi
+		break
+	done
+	while true; do
+		read -e -i "$oldPsk"     -p "Set the network's Psk (password)      : " newPsk
+		if [ -z "$newPsk" ];
+		then
+			echo -e "Error: Psk value cannot be empty."
+			echo ''
+			continue
+		fi
+		break
+	done
+	
+	sed -i -E "s|^\s*country=.*|country=$newCountry|" /etc/wpa_supplicant/wpa_supplicant.conf
+	
+	set +e #Suspend the error trap
+	sed -i -E "s|^(\s*ssid=).*|\1\"$newSsid\"|" /etc/wpa_supplicant/wpa_supplicant.conf
+	if [[ "$?" -ne 0 ]];
+	then
+		echo 'Whoops - something broke setting the SSID. Please manually edit /etc/wpa_supplicant/wpa_supplicant.conf to add the required values'
+		echo '(The most likely cause is that your SSID contains the quote (") or pipe (|) characters)'
+		echo ''
+		continue
+	fi
+	sed -i -E "s|^(\s*psk=).*|\1\"$newPsk\"|" /etc/wpa_supplicant/wpa_supplicant.conf
+	if [[ "$?" -ne 0 ]];
+	then
+		echo 'Whoops - something broke setting the Psk. Please manually edit /etc/wpa_supplicant/wpa_supplicant.conf to add the required values'
+		echo '(The most likely cause is that your Psk contains the quote (") or pipe (|) characters)'
+		echo ''
+		continue
+	fi
+	set -e #Resume the error trap
 
 	sed -i -E '/^#[^# ].*/d' /etc/dhcpcd.conf #Trim all default commented-out config lines: Match "<SINGLE-HASH><value>"
 	if ! grep -Fq "interface wlan0" "/etc/dhcpcd.conf";
@@ -744,7 +812,7 @@ END
 	
 	echo "WARNING: After the next reboot, the Pi will come up as a WiFi *client*"
 	ssid=$(sed -n -E 's/^\s*ssid="(.*)"/\1/p' /etc/wpa_supplicant/wpa_supplicant.conf)
-	echo -e "WARNING: It will attempt to connect to this/these SSIDs:\n$ssid"
+	echo -e "WARNING: It will attempt to connect to this/these SSIDs: $ssid"
 	echo "WARNING: 'sudo nano /etc/wpa_supplicant/wpa_supplicant.conf' to change"
 }
